CARBON-58: Move the PoC code into the Carbon services package We want to have all the services in the service package.
This commit also creates a script called run-provisioner-service within
the scripts folder so we can run it locally by calling the script.

Note that the specs created using swagger is into the
carbon/services/provisioner/specs folder. Whenever we need to change
the spec we have to load this file into the swagger editor, make the
changes, download the file and push the changes to Gerrit for review.

Also:
  * Updated documentation
  * Updated setuptools and tox config