CARBON-21 Create a provisioner module to wrap linchpin commands
This commit brings an updated linchpin module which contains code to
create linch-pin files, provision and teardown resources based on hosts
defined in carbon input file.

Other issues that went into this release are CARBON-21 & CARBON-22.

For initial development a forked copy of linch-pin in rywillia github,
will be used to install. Linch-pin will be installed when carbon is
installed.

Change-Id: Ideead0ca7968388bc8f7f15e0743a1392c48d380