CARBON-85 Check if the provider credentials fields are valid
This commit expands on passing provider credentials to host resources
part of CARBON-21. When each host is created it will get the credentials
for the provider declared and check the credentials fields. Ensuring
their fields match what the provider requires.

This commit also includes the following fixes:
	1. Move jinja2 declaration in requirements.txt to its associated
package.
	2. Update misleading messages.
	3. Move the code from host.desc() to host.profile(). Profile method
existed before and makes sense to have the code inside host.profile().
	4. Do not allow scenario id to be set after resources have been
instanciated.

This commit also sets the scenario id based on the scenario declared in
YAML file. In the scenario class it will update the scenario id
appending a uuid (shortned to 5 chars) to allow storing files from
provision - teardown. Easily allowing us to go back to reference data.
This is needed for linch-pin provisioner. Can be changed in the future.