CARBON-86: fixing autoscripts to work on a mac.
The problem with tox and mac is that when it needs to compile
a library during the pip install command, it does not find the
libraries under homebrew installation (usually at /usr/loca)

Therefore, we must first indetify if the tox is running on a mac,
then we set some environment variable for the openssl includes for
the cryptography package.

We also set the pkg-config variable so other libraries when compiling
can make use of this tool.