Set updated package versions.
The PR sent for taskrunner (Python3 compatibility) was merged into their
master branch. Newest version is 0.3.0. Set the new version to allow
carbon to install. Once a new taskrunner package is available on pypi,
we can use that one directly.

Changed linch-pin to download from develop branch. Once 1.0 is released,
we can change this to use the latest version.

Change-Id: I6d5d7e7675dcc82f3e4ac2fa248500bb5b1900ca