CARBON-80 Create simple validation for the Openstack provider parameters
This commit contains code that will validate all carbon resource
provider keys for Openstack providers. Validation exists for keys such
as image, flavor, network, etc. More will be added over time.

Each key validation performs the following checks:
	1. Ensures a value is provided.
	2. Ensures the value provided is the correct data type required.
	3. Ensures the value (image, flavor, etc) actually exists in Openstack
tenant.

Unit tests exist for testing the methods to validate the keys. I have
commented out parts of the tests that perform the interaction with the
tenant. Reason for this is because we need to come up with a way to
handle tenant credentials. Did not want to check in credentials in plain
text.

I also changed the name of the keypair key name. Initially it was
os_key_name but changed to os_keypair. Keypair is the name used all
throught Openstack API, CLI, UI.

One new package is required for carbon (python-openstackclient). It
currently requires a certain version which is needed by linch-pin.

Change-Id: Ia80040cbd7328a500a4c4985506c230925baa5e1