CARBON-101 Create the input resource example for openshift
This commit introduces an example scenario 'descriptor' for the
openshift use cases with carbon. It also includes the initial mandatory
and optional parameters for openshift resources.

Fixes a bug that was causing _load_resources() to raise an exception
when a scenario file did not include a carbon supported task. This
change gives users the flexibility to call certain tasks.