CARBON-125 Create the Ansible class to drive openshift client calls
This commit contains an Ansible worker class that can run modules or
playbooks and a Carbon callback class for capturing task results. Unit
tests created to test ansible worker.

Ansible package added to setup.py. It is currently set to a specific
version due to requirement by linch-pin. This can be changed moving
forward when linch-pin becomes a plugin to carbon.