Proposal for a new name
This is a propose for a new name for the provisioners we implement
that will talk directly with the provider. I think it sounds simpler
than appending carbon_ or Carbon. For example:

OpenstackProvisioner and openstack.py
OpenshiftProvisioner and openshift.py
BeakerProvisioner and beaker.py

This just reads better appending the framework name.