CARBON-159 Handle stop/remove containers for openshift provisioner
This commit enhances the openshift provisioner module to stop and
remove containers on success or failure. Currently the containers are
kept up and running. Without proper clean up, we will have lots of
containers left around. In the future we should determine to remove
the container based on clean up setting declared.

Change-Id: Ib83c134cf86dc71ff03215a494e99010fa0ec0aa