Adding data folder to save after provision files
A new variable called DATA_FOLDER was added to the configuration so
provisioners can save the hosts information captured after provisioning
is successfully executed. Based on DATA_FOLDER/host_name.yaml, this file
is now easier to be captured by other applications such as the
provisioner service.

* The configuration object is now passed to the CarbonResource objects
so they can have visibility to the configuration variables. Hence the
configuration is created and loaded by Carbon, the configuration
variables should not change.

* Updated tests to reflect the changes above

* Updated Openstack provisioner to use DATA_FOLDER.