Changes to fix data folder order issue
I reordered the way the data_folder is created by giving priority to
what the user defines when creating Carbon object.

I also removed the need to create the scenario UID in other classes
because all resources now have access to the config object, generated
when the Carbon object is created.

* I included a carbon.cfg into the tests/assets so now all tests that
need to create Carbon object will read the carbon.cfg in the test assets
and the data_folder will be created on the tests directory
(tests/.workspace). All functional tests have to be changed to set an
environment variable.