CARBON-165 Generate a status file for each carbon run
This commit brings the creation of a new scenario status file each time
carbon runs a scenario. This file is helpful when wanting to understand
any failures that may have occured or just the status of the tasks. It
will also be helpful in the future when we go to start a carbon scenario
at a certain point.

The file consists of the following data:
1. Last executed task
2. Last executed task status
3. All tasks to be run for the carbon scenario
4. List of executed tasks. This list will contain tasks executed along
with details about its resources. For example it would contain the
results from taskrunner execution. It also contains the status for the
task run (0=success, 1=failure).

Please see the CARBON-165 JIRA for attached examples of success and
failures.

I made one change to how we call taskrunner. Previously we were
passing a list of tasks to be executed. In order to ensure we have
the correct context for the task executed to be saved in our status
file. We need to send a list of one task to taskrunner. This allows
us to have the context for that specific task to easily add to status
file.

Change-Id: I742516b96c28df30fd02154c4e179de624222aa7