Expanded code coverage for carbon providers.
This commit adds unit tests to have almost 100% code coverage for all
carbon providers.

It also includes one change to the openshift labels validation method.
In Python3 you are unable to get the key:values from a dict as follows:
k, v = dict.items(). You get the following exception:

ValueError: not enough values to unpack (expected 2, got 1)

Updated the method to be compatible with Python2 and Python3.

Removes a pipeline constant from tasks package which is not used.

Change-Id: I88267fe6bcdb98a7073d9fa596656b001116b288