docs: Openstack provider documentation
This change brings the provider documentation for OpenStack. It
gives the users a guide on how to setup their credentials and
resources. Also includes a CSS file overriding sphinx_rtd_theme.
This is needed to override sphinx from displaying a scroll bar
for the tables width. It brings text wrapping to the tables.

Closes CARBON-191.

Change-Id: I170ada6955e8e9ab7feb5ed38fa8ea7c73b70921