chore: Docker controller class to have name property
This change brings the name property to the controller classes of
carbon. The docker controller was updated to pass the name of the
container to create when instantiating the class. This allows you
to not need to pass the name of the container to each method call
to perform some actions.

All provisioners using docker controller have been updated for
this change.

Closes CARBON-188.

Change-Id: Iccf0db8537cba66f4a52eda8722f24734dbc20a1