refactor: move assets folder to carbon's responsibility
The main purpose of the assets folder is to hold the assets used in the
scenario descriptor within the data folder, so whenever we run a
scenario in different tasks we know where the assets are. Instead of
giving this responsibility to the host and scenario objects, the assets
management (check & copy) should given to the main carbon object. To
find the list of assets for a given host definition, we have to ask
first to the provider what settings in a descriptor is related to an
asset. When I request host to give me the list of its assets, it will
pass the question over to the provider who will then return a list of
every file. Since the parent of each host is the scenario itself, I can
then ask the scenario to give me the list of assets related to all its
hosts.

Closes CARBON-166, CARBON-167, CARBON-168 and CARBON-169.

Change-Id: I38230a60d68cc38375744fb8a0141086bb7d983e