refactor: redefine optional and mandatory parameters for providers
It redefines the parameters for the the CarbonProvider super class.
All parameters need to be mandatory or optional. The other parameters
variables such as assets or output will be special tuples that will
say if the variable within mandatory or optional is an asset (if you
set it within the asset parameters) or output. This brings more
assurance with the parameters (you have to create validate_* for
all parameters), allowing us to extend it in the future with more
special parameters without additional development.