fix: create directory if beaker assets are within folder
when copying the beaker keytab to the /etc/beaker folder the destination
source does not create a folder if the keytab is within the folder. This
fix adds an additional step to create the folder within /etc/beaker
before copying the keytab. This way it keeps consistent with the
relativeness added to the scenario yaml.