chore: raise exception when carbon run fails
Initially when carbon.run() finished (pass or fail) it would always
create the results and status files. We now will raise a final carbon
exception if the status of carbon.run() is not 0.

Fixes the bug when running from provisioner service of saying job
provisioned when it actually failed. It is looking for an exception to
set status as failed.

Change-Id: I3915a22f7a2e6c430c0f40eb45e35f752aaa4aa5