chore: accept versions with alphanimeric
This change allows the carbon framework to be set with an alphanumeric
grupo of characters and not only numbers. In addition to this change
there is also a script file that can change the version for the
framework to the desired value or to a continuous development version
format as per PEP 440.

 $(root) contrib/change_version.sh 0.9.1

It will change all necessary files to have version 0.9.1. This is
helpful when generating automated deployments. If no version number
is given, then it will add the following:

 <major>.<minor>.dev<timestamp>