chore: add line break when creating ssh public key
This commit adds a new line to the end of the ssh public key file
created. If a user injects an additional key into a metal machine
provisioned by the service. SSH connections will fail on the next logon.
Adding the new line at the end of the created public key resolves this
issue.

Closes CARBON-217