refactor: remove config and compat from code base
we have an exact copy of config.py from Flask. Going upstream
with this file would have to pass through complicated dual-licensing
process that we can avoid by just adding Flask as a dependency and
use Flask's config module. Also, we are just using a couple of types
from the compat library, so we have to keep it to what need only.

Change-Id: I146cac34da0fa9068cf2968a39d3b5184274b0a2