feat(signals): add signal generators
We want to emit a signal for all major tasks within the framework
modules. This is useful for external apps that are integrating the
framework to gather data while the framework is executed. These
signals will also support the development of the notifications
feature that will be added in the future.

Closes CARBON-255

Change-Id: I2660e9fd669c61ff3d00ede98865b18a4b676e35