fix: remove positional parameter from signal.send
The only parameter that can be passed without associating with
a argument is the first one, which is considered "sender" by the
blinker manual. The other parameters send through the creation of
the signals need to be associated to an argument. This change also
removes a test code that was added to the carbon.cli module.

Related to CARBON-255

Change-Id: I8cdd71c13a17ae31eee1f14759f21a05e13f969f