fix: add retry function to docker_inventory script
The docker_inventory script should be resilient in case there
are modification in the docker hosts list while the script builds
up a list of docker images. Before, the script basically builds a
list of docker images currently running in the host and pass it
to Ansible so commands can be sent to the correct docker running
instance. If there are two or more jobs running concurrently the
list taken by one job may have a deleted host, which may have been
deleted by other job, that could cause an exception while building
the json description for Ansible, saying the image is not valid.
This fix adds a retry that if an exception is thorwn during the
JSON creation, it will retry to build the list again. The number
of retries are configurable, and the default is 3.

Relates to and closes CARBON-253

Change-Id: I052bff2433b4878ac23539ef2e3207cfa2471c76