fix: exit carbon if any task raises an exception
This commit fixes a bug that was introduced when the feature for
concurrent runs was added. Blaster library allows the user to define
if they want to raise an exception if any task failed. Originally this
was set to false so we never were exiting out properly. Switched to
always raise on failure so we can exit properly again.

Closes CARBON-268

Change-Id: Ia333ead11fe8bd5e45aed74735577cfde7b8eb22