fix: catch all exceptions that may be raised by openstack client
This commit fixes a bug that leaves carbon hanging if invalid
authentication details are given. We are seeing this now with the
tasks running concurrently. Part of that feature, after each
openstack client call, we need to unitialize the openstack client
object. Why is this needed? Each blaster process adds task data
(which may include objects) into the queue to be sent back to the main
process at the end of execution. In this case the openstack client
objects are not pickable which causes the blaster to hang when trying
to finish the processes (add data to queue). The current fix for this
is to uninitialize the object before adding to the queue. With this and
catching all exceptions in place. A carbon run with invalid
authentication will exit cleanly.