fix: unable to locate kickstart file during beaker provisioning
This commit fixes a bug that was introduced by a new change of not
copying files into the container. Files are now available to the
container by a mounted volume. Originally the kickstart was needed to
be at tmp folder. Now it is stored at the mounted volume. Fixed the
kickstart path to point to new location.

Closes CARBON-292