feat: carbon's own time mixin class
This commit brings a new time mixin class for all carbon core classes.
This class provides the ability to save a start and end time. It will
calculate the time delta. It can be used to get the total duration it
took a task or the entire carbon run to complete.

Closes CARBON-271