fix: always save the beaker job id in carbon results file
This commit fixes a bug found when a provision job fails, it does not
save the beaker job id. This would result in not allow the user be
able to cancel their job. This patch fixes this by updating the host
objects returned from the provision run process. Resulting in the user
being able to delete their resource.

Closes CARBON-283