chore: provide support for ansible greater than 2.4
This commit brings compatibility for carbon with ansible > 2.4 or < 2.3.
The newest ansible release brings an updated Inventory system which
changes the class name and module locations for the Python API. Updated
our controller module to be backwards compatible with any ansible 2.x
version.

In a future release we can depreciate ansible support < 2.4. 2.4 brings
more support for Python 3.

Closes CARBON-294