chore: convert function checking git urls to be a generic url checker
This commit revises the existing function checking if a git url is
valid. It was using subprocess library to run a git command to validate.
The new function is generic which accepts any url and will validate it.
The function still returns the same boolean values as before. This
allows other carbon components to utilize this function.