chore: split requirements.txt into multiple requirements.txt files
This commit splits the main requirements.txt file into smaller
requirements files. Why is this needed? This will provide carbon
developers with a better way to visualize and understand which area of
carbon those specific packages are needed for. If you only want to
install packages to run unit tests, you would just need to install the
tests.txt file. Removing the need to install other packages
for documentation or a new release.

Change-Id: Ic07b42e61d998a2ed63f870e20c376f6394a8436