chore: define whether carbon tasks runs concurrently or not
Previously each carbon task would run concurrently. When we get to
tasks such as orchestrate, these may not be able to run concurrently.
Example is one action is dependent on another. For these cases it
should run sequentially. This commit is the start of customizing which
mode a task could execute in. Everything will be run concurrently
besides orchestrate and execute tasks.

Closes CARBON-328.