feat: initial base structure for ansible orchestrator
This commit puts in place the initial foundation for processing ansible
actions. It provides the ability to generate the inventory file with
all the correct host sections and master group and execute the action
given.

Many TODO's are still open such as determining location where playbook
is.

Part of CARBON-326.