feat: implementation for locating files for ansible actions
This commit adds the implementation to locate the ansible action
(which is the playbook). It will look for the file in the current
working directory or a user defined directory. If the same file exists
in both directories, the last one located will be used. The user defined
directory will be the assets directory path. Users just need to put
there ansible orchestrator files in that directory and carbon will take
it from there.

For more details please see CARBON-336.

Change-Id: Ie090615da10a32285a035458d83b504e48af860a