fix: resolve mapping string hosts to their actual host object
This commit fixes a bug found where the orchestrate task was unable to
map the actions given hosts to the actual carbon host objects. Carbon
by default will replace any host name with special characters. This
was causing the mapping to fail to the actual actions specified. Before
mapping in orchestrate, we will perform the same filter to ensure the
host names are correct.

Closes CARBON-340.