chore: append orchestrator name to assets if directory exists
Every orchestrator will have a directory where it can store its
associated files. If the user wants to define custom files for the
orchestrator to use, they need to create a folder in their assets
path with the orchestrator name. Previously carbon was failing if this
folder did not exists (thinking it was required). This is optional and
if not found it will not be appended to the list of assets to copy.