feat: create results folder per execution
This commit allows carbon to create a hidden results folder within the
root of the data folder. This folder will contain the most recent
status and results files for the given carbon execution. It provides
users with the ability to run carbon tasks individually and use the
updated scenario file without needing to search for the unique data
folder created. It allows them to point to the common folder.