chore: move carbon source folder to root of project
Before carbon had a src folder which contained the carbon
source folder. Developers using visual studio code with the
vscode-python plugin, will run into issues when trying to
go to library import definitions. The vscode-python plugin
for visual studio code has a list of known standard libraries.
This library contains a library named Carbon. This was causing
a conflict not allowing you to lookup the actual class/function,
etc for the carbon project.

To allow this compatibility across multiple IDE's. We can move
the source folder for carbon to the root of the project.