chore: allow inventory to contain reference to all scenario hosts
Previously the ansible orchestrator would create a unique inventory for
the given action with its defined hosts. In some cases an action may
need to access information for other hosts in the given scenario which
are not defined in the given action. i.e. running a play on a test
driver system and you need information about a test client system. This
commit modifies how the inventory is created per action. It now has
references for all hosts and the group which contains only the defined
hosts the action should run against. This allows playbooks to gather
facts for all hosts and use them in another play within a playbook.
See orchetrator/ansible/access_multiple_hosts.yml for an example.

Change-Id: I78655ae74fab247c72bcd39646c6fc6875bde673