chore: modify inventory to set alias for a host when creating host group
Certain cases you may not have access to dns for your set of systems.
This makes it difficult for you to determine a host to access from
hostvars within a play since it will use the IP address. Which would
require logic to determine if its the correct host to use. Ansible
provides the ability to set a key named ansible_host which aligns with
the alias defined for the host. This way the play hostvars will have
the alias of your system easily allowing you to reference other hosts
within a play or multiple plays.

This commit includes the required code changes and updates to one of
the playbooks demonstrating this change.