chore: allow users to set beaker server url
By default users did not need to declare the beaker server url. It
would use the setting defined in carbon constants. This commit gives
the ability to have the user define this. In a future release, this
will become a mandatory beaker credential key to be set. It is needed
to make the framework capable of moving upstream.

Change-Id: I00606202f303f3d0c4a34828703f675499424400