refactor: remove the option for selecting log type
Previously carbon provided the ability to either log the output
to the console stdout or a file. Logging to a file was necessary
when using carbon framework in the service model. It was decided
that both should be done. With this change, carbon will always
log to the console and to file for archiving purposes.

Change-Id: Ia53f4367c68250058823efc2b82751a8a27a3908