documentation restructure
This commit restructures the documentation for carbon. Previously it
was structured in the service model. This update puts the structure in
place for how users use carbon, setup their input files, how they can
contribute and details about the framework architecture.

Note: Some sections of the documentations still needs content.

Change-Id: I7af991da86b33dee039055db2644d0df10c28ca4