add ability for carbon to download and consume ansible roles
Previously any ansible action processed by carbon was unable to use
roles. This patch provides the ability for any action to use roles
provided from galaxy. This was needed to allow carbon to have the
ability to use roles from oasis-roles. The action definition changed by
removing the vars key and replacing with ansible_options and a new key
ansible_galaxy_options. Providing the action to have more flexibility
in which options it can consume.

Please see the following JIRA issues below for more details.

Associated with CID-3407 and CID-3408

Change-Id: I276ba1d7820ae26ce686973125ec4dfdfbb2fe6c