add a description key for all carbon resources
Every resource which makes up the carbon compound has a set purpose.
Currently from the scenario descriptor it is hard to tell exactly what
each resources purpose is. This patch allows each resource to set a
description. This description will provide an easier way to understand
what exactly the scenario is accomplishing at that given time. It also
opens the ability for carbon to use that description at a later point
in time (i.e. logging and so on). Last point, this is an optional key.

Associated with CID-3406

Change-Id: I535be8a91c0929d014ae772d29cfacc5ef54f86b