ability for orchestrate tasks to easily define all hosts to run against
Previously if you wanted to run a orchestrate task against all the hosts
in the scenario, you would need to type everyone. This gets tedious
quickly when you have a large amount of hosts and tasks. This patch
gives the ability for a task to say all hosts. Removing the need to
declare every hosts.

Change-Id: I903b0a5a8323dae467c7ff8f247c985188a2535b