only attempt to apply task defined extra vars when they exists
Previously every time carbon would attempt to update the extra vars to
be passed to ansible with the extra vars defined by the orchestrate task.
If no extra vars are defined, carbon would fail since the default value
is None. This change makes the ansible_options and ansible_galaxy_options
have default values of a dictionary if not defined. Then the ansible
orchestrator class will check if extra vars is set first before updating
the extra vars to be passed to ansible playbook call.

Change-Id: I3d58bd9376f675ced76e8eca15b637e13b259038