rename carbons command line interface click group
Initially the cli of carbon would create a click group named cli to
hold all carbons sub-commands. While testing carbon on centos, it was
observed that carbon would fail with an error saying group object has
no attribute options. When ansible is loaded it tries to load cli
object from __main__. Since one was already created by carbon ansible
loads it but then fails since that cli is missing the options attribute.
This for some reason is not exposed on other operating systems. To
handle this error and resolve any future problems, we should not mask
the same name in use by ansible. This patch changes carbons cli group
to carbon from cli.

Change-Id: I613de4ddb41e5925cf5d3d4a7b23ca84eee1f361