remove support of assets path option
Previously carbon had the model of any files needed by the scenario
would need to be in the assets path. This model does not scale very
well because it limits a scenario to put files in certain locations.
The assets path was depreciated and replaced by a workspace option.
Now you have the ability to set the workspace to a directory where
carbon can use any files within for the scenario. This then provides
the scenario with the flexibility to put files in any style they wish.

This patch also changed how you define an ansible orchestrate task.
Previously you would just set the file name including a relative path
to the file within the workspace (if needed). It was decided to make it
more clear to have user set the file extension instead of looking it up.
This will make carbon more standard as we implement more orchestrators.

Associated with CID-3451

Change-Id: I04a653bb414ceac7dfe4bda5d03cd4e2c74b159b