restructure how configuration settings are defined
This change reformats the configuration file to be ini
formatted. It allows to separate settings into sections
for easy readability. A defaults section can exists for
any common config, credentials sections per provider
related config and orchestrator sections for related
config. This also removes the need for carbon to require
flask as a package dependency.

Aligns with CID-3460.

Change-Id: I9a1f97a3fa97718757d116b56deb65cd378b54e5