Implementation for running user defined scripts.
The implementation for running user defined scripts was added to the
ansible orchestrator.  There is an additional option key
(ansible_script), which should be set to true, location of the script
from the workspace in name, and extra_args passed as an addition option
in ansible_options.

The following is an example of usage:

orchestrate:
- name: scripts/create_dir.sh
  description: create a remote dir
  ansible_options:
    remote_user: cloud-user
    extra_args: -c -e 234ansible_script: True
  hosts: <name of machine or localhost>
  orchestrator: ansible

Closes: CID-3396