reduce the size of the run method for the runner executor
This commit breaks down the run method into smaller methods which it
will call. Each method will be called if the supplied attribute exists
and is not empty. With smaller methods for each given action, it helps
make the code easier to maintain and easier to write test cases.

Relates to CID-3546.

Change-Id: Ie6eee91926242f4fd9a4e03af90031f405d4c4dd