flag to control test execution status
When a test command fails to run successfully, carbon raises a custom
exception to tell it to attempt to fetch test artifacts that may have
been produced during some test execution. Once the artifacts are saved,
carbon would execute successfully when tests execution failed. This
commit adds a new status attribute when is modified when a command fails
to run. Once archiving is complete it will fail carbon based on the
status (0/1).

Change-Id: Ice7ccc598c31671856dd014373a425aa68b18571