adds a class to perform data injection into strings
This commit brings a new class that allows you to inject data into
strings. This was needed for carbons execute task. Some test types
require machine ip addresses as options to the test framework to drive
the tests (i.e. restraint). This commit allows your carbon definition
file to define a variable such as { host01.ip_address } and this new
class will evaluate the string and lookup the host01 resource object
and replace it with the correct value for the ip_address attribute. This
works for any other attribute the host has to offer. It is great if you
have data defined in the metadata section. (i.e. { host01.metadata.key1 }

Change-Id: Ia80a6389e030d4d4f84832bc47eee23b08d520d6