do not add credentials to updated input file if set in config
Previously if you were to set your credentials within a carbon config
file, it would write the credentials to the updated input scenario file.
When you set your credentials within a carbon config, they should not
be wrote to the updated input scenario file. The only time they should
is if they are defined in the input file to start. This commit resolves
this bug and will properly write the credentials as needed.

Closes CID-3616

Change-Id: Ie7d6c6cd36453cdfca72598ee999b78110ef14a1