resolves a bug causing tests to not run on the correct host defined
This commit resolves a bug found when carbon is called running all
tasks where the execute task is unable to run the test against the
defined host. The problem was the ansible inventory directory contained
multiple inventories with the same group but each group had different
host entries. Ansible was selecting the incorrect one that the test
needed to run against. This commit resolves this issue by cleaning up
the unique inventory per ansible call. This resolves CID-3617.

This commit also resolves CID-3534. At the end of the carbon run, all
files archived in the unique data folder for that run are also copied
to the results folder. This allows for easy access to the latest runs
results.

Change-Id: I728ee65ec3e77efc3858c3a96710037f88a37f51