refactor how hosts are defined
This commit brings a new approach to how hosts are defined for a
carbon scenario. Previously each host had a mapping to a provider
which had various keys that could be set. These keys required a
provider prefix model. To simplify this approach, the provider model
still exists but is scaled down. This means the provider classes will
only contain attributes for the various keys a provider can have.
All provider validation should be handled by the provisioner tool
of choice. It also removes the need for the provider prefix. Now
under one key 'provider' all necessary required and optional keys
will reside there. This should hopefully make it transparent between
the various provisioning tools that teams used today. Making it easy
for them to plug into carbon. The provisioner classes would have
access to a dictionary 'provider_params' which has all the data
needed for that given provider to fulfill the request.

This change is needed prior to creating an example blueprint class
for adding additional provisioners.

Change-Id: I4986084def68b57b2f8f240b54632edcb1f9a7ae