skip validation for static hosts
A bug was found where host validation would fail for static hosts. The
reason was due to static hosts do not have an associated provider. We
need to skip validation for static hosts. This commit sets a flag to
control whether the host is static or needs to be provisioned.

Change-Id: I3c4cc96e4f3231109143943bea9b9fadc53f9581