fix beaker provision error when no ssh key is defined to be injected
A bug was found where if you did not set the ssh_key for the host
resource, the provisioner class would fail due to it was unable to
find the key within the provider parameters. This commit resolves the
issue and will only attempt to inject the ssh key if the key is
properly set.

Change-Id: Ied77f1373bba827f9f247db68f10a9e7997c7a88