modify the tox doc environment to clone the examples repo
Carbon now has an external repository where all its examples are stored.
Sphinx currently does not have a way to do a literal include for a url.
This change modifies the tox doc environment to clone the examples
repo and put it in the docs directory. This then allows you to
include these examples within the carbon doc files.

i.e.

.. literalinclude:: .examples/README.rst

Change-Id: I0c934697c1b301a6656f8c1f9b6eda10320ee502