do not jump out of the loop if the host is localhost
In order for ansible to successfully allow playbooks or ad hoc
commands to run against localhost. The group in the inventory file
needs to have the ansible_connection=local set. Previously carbon was
skipping setting ansible params for localhost connections. This commit
removes the jumping out of the loop to allow ansible params to be set
for localhost.

Change-Id: I5b61217bb5a1943ad6be7260951d9848b5b3f1fd