remove applying filter to host resource name
Carbon should not apply filtering on the host
resource name the user defines. This name is referenced
by other carbon tasks such as orchestrate & execute.
If you apply filtering on the name, the updated
descriptor file (results.yml) will not match up with
what the user initially defined. This change is to remove
filtering and only apply filtering on the hostname given
to the provisioner.

This also resolves bug: CID-3655

Change-Id: If2bb4c62b798709d36092cd6aa9dc574502116bc