Only install rsync and git when required
This change runs the tasks for installing git and rsync packages
only when they are not present. This is required to allow running
carbon as a non-privileged user inside a container where running
tasks as root is not an option. Instead packages will be preinstalled
and ready to be used by used carbon.

Change-Id: I420bcf8e419bf71904d6d184560092a0e121f0f8